[build-system] Generic paths, attempts to build the x86 libs
* add spec.template and put packageinfo in correct location
* correction of types, serializer for `size_t`

+++ src/utils.cpp
@@ -608,7 +608,7 @@ void utils::writeBlock(const std::string &regName, const uint32_t *values, const
   }
 }
 
-void utils::writeBlock(const uint32_t &regAddr, const uint32_t *values, const size_t &size, const uint32_t &offset)
+void utils::writeBlock(const uint32_t &regAddr, const uint32_t *values, const uint32_t &size, const uint32_t &offset)
 {
   // This function doesn't make sense with an offset, why would we specify an offset when accessing by register address?
   // Maybe just to do validation checks on the size?
